Add function to compute how many first descendants
In reductions it is necessary to figure out how many contributions should be
made at various levels of the virtual spanning tree across nodes. This is done
by running a predicate over the different element IDs on the different
processes, and the general predicate to the count_first_descendants function
will be used for this.