Improve Charm++ serialization interoperability
We no longer have a generic catch-all operator|(PUP::er&, ...). This caused
infinite recursion that was really annoying to debug. Instead we provide all the
necessary STL specialization and generic overloads for fundamental and enum
types, as well as for classes that define a pup(PUP::er&) member
function. This should cover all use cases that I (Nils Deppe) currently think
are reasonable. Charm++ has their own array and pointer types, but I'm not
convinced those are good to use.